fix(test): four pre-existing CI flakes blocking #471
Buildkite has been failing on main for a while; the failures aren't
caused by #471 but they masked it. Each fix below was verified by
running the offending test file in isolation, then by full `npm test`.

1. **vault broker client**: bun (1.3.x) emits the `connect ENOENT`
   synchronously from inside `net.createConnection({path})` when the
   socket file is absent — before the `.on('error', ...)` listener
   attached on the next line could catch it. Switched to
   `new net.Socket()` + listeners + `.connect()` so handlers are wired
   up first under both runtimes. Repro:
   `bun test src/vault/resolver-via-broker.test.ts`.

2. **scaffold A4a + A4b**: both tests wrote a temp `.md.hbs` template
   into the shared `profiles/default/workspace/` and cleaned up via
   try/finally. Vitest runs files in parallel — when scaffold.test.ts
   was in its cleanup window, scaffold.hot-reload-stable.test.ts
   walking the same dir would `readdirSync` the file then ENOENT
   on `statSync`. Now both tests build a sibling temp profile
   (`profiles/__test_a4{a,b}_<pid>_<ts>`) by copying default; the
   `__` prefix keeps it out of `listAvailableProfiles`. Added a
   matching `.gitignore` entry as a belt-and-braces orphan-cleanup
   safety net.

3. **silent-reply-guard**: test imported from `../server.js`, which
   has top-level side effects (env load, `TELEGRAM_BOT_TOKEN` check,
   `process.exit(1)` if missing, session-tail spawn) that killed the
   bun-test runner mid-batch. Extracted `guardSilentReply` +
   `isSilentReplyMarker` to `telegram-plugin/silent-reply.ts`; server.ts
   re-exports them so internal call sites are untouched. Test now
   imports the helper module directly with no side effects.

4. **subagent-tracker-posttool**: the "completed with result_summary"
   test passed `run_in_background: true`, but PostToolUse's correct
   semantics for background agents is to bump `last_activity_at` and
   capture `result_summary` only — status stays 'running' until the
   watcher's recordSubagentEnd fires on turn_end. Test was asserting
   foreground-completion semantics on a background row. Flipped to
   `run_in_background: false` (matches the test name) and documented
   the foreground/background split in the test comment.

After this commit: vitest run is clean, bun-test sweep is clean,
`npm test` exits 0 (4109 vitest + 285 bun = 4394 tests passing).
